Rework `FineFormIter`
This should make it easier to understand the structure of the iterator
and to make it easier to output only certain levels or only the basic
normal forms.

Compatibility is preserved. I.e. the iterator produces precisely the
same output. It is marked in the code, which parts could be removed to
still enumerate the all formulas, but which would break backwards
compatibility.